# Optimize GQA/MQA